Refactor handoff output writers into SRP modules
